Trawl a local CRAN archive and extract statistics from all packages

Trawl a local CRAN archive and extract statistics from all packages

pkgstats_from_archive(
  path,
  archive = TRUE,
  prev_results = NULL,
  results_file = NULL,
  chunk_size = 1000L,
  save_full = FALSE,
  results_path = tempdir()
)

Arguments

path
Path to local CRAN archive

archive
If TRUE, extract statistics for all packages in the /Archive sub-directory, otherwise only statistics for main tarballs directory (that is, current packages only).

prev_results
Result of previous call to this function, if available. Submitting previous results will ensure that only newer packages not present in previous result will be analysed, with new results simply appended to previous results. This parameter can also specify a file to be read with readRDS().

results_file
Can be used to specify the name or full path of a .Rds file to which results should be saved once they have been generated. The '.Rds' extension will be automatically appended, and any other extensions will be ignored.

chunk_size
Divide large archive trawl into chunks of this size, and save intermediate results to local files. These intermediate files can be combined to generate a single prev_results file, to enable jobs to be stopped and re-started without having to recalculate all results. These files will be named pkgstats-results-N.Rds, where "N" incrementally numbers each file.

save_full
If TRUE, full pkgstats results are saved for each package to files in results_path.

results_path
Path to save intermediate files generated by the chunk_size parameter described above.

Value

A data.frame object with one row for each package containing summary statistics generated from the pkgstats_summary function.

Note

Each analysis in an archive trawl spawns several unsupervised processes, preventing the trawl from running in parallel. Accurate results can only be guaranteed by running this function as a single process.